pyproject.toml
[project]
name = "gh-release-notifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx[http2]>=0.28.1",
    "pydantic-settings[yaml]>=2.11.0",
    "pygithub>=2.8.1",
]
